feat: Organize public API headers in include/nhssta/ (Issue #44)
- Move public API headers to include/nhssta/:
  - exception.hpp -> include/nhssta/exception.hpp
  - ssta_results.hpp -> include/nhssta/ssta_results.hpp
  - ssta.hpp -> include/nhssta/ssta.hpp
- Update all include statements to use <nhssta/...> format
- Update Makefile to add -I../include for include path
- Add comprehensive tests for public API headers
- Internal implementation headers remain in src/

All 228 tests pass (4 new tests added).
