+++ include/nhssta/exception.hpp
