+++ src/random_variable.hpp
@@ -9,7 +9,7 @@
 #include <type_traits>
 
 #include "smart_ptr.hpp"
-#include "exception.hpp"
+#include <nhssta/exception.hpp>
 
 namespace RandomVariable {
