+++ test/test_ssta_unit.cpp
@@ -3,8 +3,8 @@
 // Phase 2: Comprehensive unit test coverage for Ssta class
 
 #include <gtest/gtest.h>
-#include "../src/ssta.hpp"
-#include "../src/ssta_results.hpp"
+#include <nhssta/ssta.hpp>
+#include <nhssta/ssta_results.hpp>
 #include <fstream>
 #include <sstream>
 #include <iostream>
